feat/add mantenimineto de sesion (#80)
