+++ client/src/index.css
@@ -4,7 +4,7 @@
 
 body
 {
-  overflow-x: hidden;
+
   
 }
